+++ packages/web/src/controls/TextInput.tsx
@@ -185,8 +185,8 @@ export const TextInput = memo(
     const mergedProps = useComponentConfig('TextInput', _props);
     const {
       label,
-      labelFont = 'label1',
-      labelColor = 'fg',
+      labelFont = 'label2',
+      labelColor = 'fgMuted',
       accessibilityLabel,
       helperText = '',
       variant = 'foregroundMuted',